+++ tests/filecheck/frontend/dialects/arith.py
@@ -1,9 +1,9 @@
 # RUN: python %s | filecheck %s
 
-from xdsl.frontend.context import CodeContext
-from xdsl.frontend.dialects.builtin import f16, f32, f64, i1, i32, i64
-from xdsl.frontend.exception import CodeGenerationException
-from xdsl.frontend.program import FrontendProgram
+from xdsl.frontend.pyast.context import CodeContext
+from xdsl.frontend.pyast.dialects.builtin import f16, f32, f64, i1, i32, i64
+from xdsl.frontend.pyast.exception import CodeGenerationException
+from xdsl.frontend.pyast.program import FrontendProgram
 
 p = FrontendProgram()
 with CodeContext(p):
