+++ tests/filecheck/frontend/dialects/builtin.py
@@ -1,8 +1,8 @@
 # RUN: python %s | filecheck %s
 
 
-from xdsl.frontend.context import CodeContext
-from xdsl.frontend.dialects.builtin import (
+from xdsl.frontend.pyast.context import CodeContext
+from xdsl.frontend.pyast.dialects.builtin import (
     f16,
     f32,
     f64,
@@ -15,7 +15,7 @@ from xdsl.frontend.dialects.builtin import (
     ui32,
     ui64,
 )
-from xdsl.frontend.program import FrontendProgram
+from xdsl.frontend.pyast.program import FrontendProgram
 
 p = FrontendProgram()
 with CodeContext(p):
