+++ tests/test_frontend_type_conversion.py
@@ -7,11 +7,11 @@ from typing import Any, Generic, Literal, TypeAlias, TypeVar
 
 import pytest
 
-from xdsl.frontend.dialects.builtin import (
+from xdsl.frontend.pyast.dialects.builtin import (
     _FrontendType,  # pyright: ignore[reportPrivateUsage]
 )
-from xdsl.frontend.exception import CodeGenerationException
-from xdsl.frontend.type_conversion import TypeConverter
+from xdsl.frontend.pyast.exception import CodeGenerationException
+from xdsl.frontend.pyast.type_conversion import TypeConverter
 from xdsl.ir import ParametrizedAttribute
 from xdsl.irdl import irdl_attr_definition
