+++ xdsl/frontend/pyast/dialects/arith.py
@@ -2,7 +2,7 @@ from collections.abc import Callable
 from typing import TypeVar
 
 import xdsl.dialects.arith as arith
-from xdsl.frontend.dialects.builtin import f16, f32, f64, i1, i32, i64, index
+from xdsl.frontend.pyast.dialects.builtin import f16, f32, f64, i1, i32, i64, index
 from xdsl.ir import Operation
 
 _Int = TypeVar("_Int", bound=index | i1 | i32 | i64)
